:art:
The Unet class can be written in the main module file, this makes it cleaner and more organized for the individual implementations of each model
