+++ segnet/models/unet.py
@@ -1,7 +1,7 @@
 from tensorflow import keras as K
 
 
-def _simple_unet(input_size, conv):
+def simple_unet(input_size, conv):
     """
     Implementation of the U-Net model, using Concatenation instead of
     crop and place for the semantic gap.
@@ -95,31 +95,4 @@ def _unet_encoder(x, pool, conv):
     some_layer = K.layers.MaxPooling2D(pool_size=pool)(some_layer)
 
     return some_layer
-
-
-class Unet:
-    def __init__(self, input_size, variant=None):
-        self._input_size = input_size
-        self._activation = None
-        self._padding = None
-        self._pool = None
-        self._l1_reg = 0.0
-        self._l2_reg = 0.0
-        self._seg_model = None
-
-        if variant is "simple":
-            self._activation = "relu"
-            self._padding = "same"
-            self._simple_init()
-
-    @property
-    def model(self):
-        return self._seg_model
-
-    def _simple_init(self):
-
-        options = {"activation": self._activation, "padding": self._padding}
-
-        self._seg_model = _simple_unet(self._input_size, options)
-
-    
+
